Fix bugs and organize typedefs
