+++ server/scripts/genTeamInfo.ts
@@ -1,51 +1,25 @@
 import fetch from 'node-fetch';
 import { dotenvLoad } from 'dotenv-mono';
 import fs from 'fs';
+import { TeamData, TeamInfo } from 'requests'
 
-interface TeamData {
-    [teamNumber: string]: {
-        primaryHex: string;
-        secondaryHex: string;
-        verified: boolean;
-        avatar?: string;
-        info: TeamInfo | {
-            Error: string;
-        };
-    };
-};
+interface ColorData {
+    primaryHex: string;
+    secondaryHex: string;
+    verified: boolean;
+}
 
-interface TeamInfo {
-    address: null;
-    city: string | null;
-    country: string | null;
-    gmaps_place_id: null;
-    gmaps_url: null;
-    home_championship: Record<string, string> | null;
-    key: string;
-    lat: null;
-    lng: null;
-    location_name: null;
-    motto: null;
-    name: string;
-    nickname: string;
-    postal_code: string;
-    rookie_year: number;
-    school_name: string;
-    state_prov: string;
-    team_number: number;
-    website: string | null;
-};
-
-interface AvatarData {
+// avatar info is a list of multiple
+type AvatarData = {
     foreign_key: string;
     preferred: boolean;
-    type: string;
+    type: 'youtube' | 'cdphotothread' | 'imgur' | 'facebook-profile' | 'youtube-channel' | 'twitter-profile' | 'github-profile' | 'instagram-profile' | 'periscope-profile' | 'gitlab-profile' | 'grabcad' | 'instagram-image' | 'external-link' | 'avatar';
     details: {
       base64Image: string;
     };
     direct_url: string;
     view_url: string;
-  }
+}[]
 
 
 dotenvLoad({ path: '.env' });
@@ -53,7 +27,7 @@ dotenvLoad({ path: '.env.local' });
 const apiKey = process.env.API_KEY!;
 
 // Get every team number in the output_analysis.json file
-const teams = JSON.parse(fs.readFileSync('static/output_analysis.json', 'utf-8')).map((e: {teamNumber: number}) => e.teamNumber);
+const teams = JSON.parse(fs.readFileSync('static/output_analysis.json', 'utf-8')).map((e: {teamNumber: number}) => e.teamNumber) as number[];
 
 // Create a teamColors object to store the color for each team
 const teamInfo: TeamData = {};
@@ -63,7 +37,7 @@ console.log("Getting team colors...");
 // Get the color for each team. If it returns a 404, default to gray
 for (const team of teams) {
     const color = await fetch(`https://api.frc-colors.com/v1/team/${team}`);
-    const colorJson = await color.json() as TeamData;
+    const colorJson = await color.json() as ColorData;
     if (color.status === 404) {
         teamInfo[team] = {
             primaryHex: '#7f7f7f',
@@ -73,10 +47,10 @@ for (const team of teams) {
         };
     } else {
         teamInfo[team] = {
-            primaryHex: colorJson[team].primaryHex,
-            secondaryHex: colorJson[team].secondaryHex,
-            verified: colorJson[team].verified,
-            info: colorJson[team].info,
+            primaryHex: colorJson.primaryHex,
+            secondaryHex: colorJson.secondaryHex,
+            verified: colorJson.verified,
+            info: {Error: "Info not loaded"},
         };
     }
 }
@@ -85,14 +59,15 @@ console.log("Getting team avatars...");
 
 // Get the avatar for each team.
 for (const team of teams) {
-    const avatar = await fetch(`https://www.thebluealliance.com/api/v3/team/${'frc' + team}/media/2024`, {
+    const avatar = await fetch(`https://www.thebluealliance.com/api/v3/team/frc${team}/media/2024`, {
         headers: {
             'X-TBA-Auth-Key': apiKey,
         },
     });
     const avatarJson = await avatar.json() as AvatarData;
     if (avatar.status !== 404) {
-        teamInfo[team].avatar = avatarJson.details.base64Image;
+        // Some teams have no media results
+        teamInfo[team].avatar = avatarJson.find(e => e.type === 'avatar')?.details.base64Image;
     }
 }
 
@@ -100,7 +75,7 @@ console.log("Getting team info...");
 
 // Get the information for each team
 for (const team of teams) {
-    const info = await fetch(`https://www.thebluealliance.com/api/v3/team/${'frc' + team}`, {
+    const info = await fetch(`https://www.thebluealliance.com/api/v3/team/frc${team}`, {
         headers: {
             'X-TBA-Auth-Key': apiKey,
         },
